RF: Strip AnnexRepo._set_direct_mode() from class and make test helper
That is its only purpose. Any chance to reduce clutter in this monster
class should be used.
